PagesEditor: split WorkflowHeader into WorkflowTopLinks and WorkflowTabs
