+++ app/pages/lab-pages-editor/PagesEditor.jsx
@@ -12,7 +12,8 @@ import { StrictMode, useState } from 'react';
 import PropTypes from 'prop-types';
 
 import DataManager from './DataManager.jsx';
-import WorkflowHeader from './components/WorkflowHeader.jsx';
+import WorkflowTabs from './components/WorkflowTabs.jsx';
+import WorkflowTopLinks from './components/WorkflowTopLinks.jsx';
 import TasksPage from './components/TasksPage';
 import WorkflowSettingsPage from './components/WorkflowSettingsPage';
 import strings from './strings.json';
@@ -31,11 +32,11 @@ function PagesEditor({ params }) {
   const tabs = [
     {
       id: 'pages-editor_workflow-header-tab-button_task',
-      label: strings.PagesEditor.components.WorkflowHeader.tasks,
+      label: strings.PagesEditor.components.WorkflowTabs.tasks,
       targetPanel: 'pages-editor_tab-panel_task'
     }, {
       id: 'pages-editor_workflow-header-tab-button_settings',
-      label: strings.PagesEditor.components.WorkflowHeader.workflow_settings,
+      label: strings.PagesEditor.components.WorkflowTabs.workflow_settings,
       targetPanel: 'pages-editor_tab-panel_settings'
     }
   ];
@@ -48,12 +49,20 @@ function PagesEditor({ params }) {
           projectId={projectId}
           workflowId={workflowId}
         >
-          <WorkflowHeader
+          <WorkflowTopLinks
             currentTab={currentTab}
             projectId={projectId}
             setCurrentTab={setCurrentTab}
             tabs={tabs}
           />
+          {/*
+          <WorkflowTabs
+            currentTab={currentTab}
+            projectId={projectId}
+            setCurrentTab={setCurrentTab}
+            tabs={tabs}
+          />
+          */}
           {(currentTab === 0) && (
             <div
               aria-labelledby={tabs[0].id}
